KEY-5 Extract models to separate file
