+++ src/models.ts
@@ -1,0 +1,7 @@
+export type Rectangle = DOMRect;
+
+export type Highlight = {
+  clickable: HTMLElement;
+  element: HTMLElement;
+  rect: Rectangle;
+};
